add: authorization

+++ app/Repositories/Eloquent/ReplySupportRepository.php
@@ -4,6 +4,7 @@ namespace App\Repositories\Eloquent;
 use App\DTO\Replies\CreateReplyDTO;
 use App\Repositories\Contracts\ReplyRepositoryInterface;
 use App\Models\ReplySupport as ModelSupport;
+use Illuminate\Support\Facades\Gate;
 use stdClass;
 
 class ReplySupportRepository implements ReplyRepositoryInterface
@@ -41,6 +42,12 @@ class ReplySupportRepository implements ReplyRepositoryInterface
         {
             return false;
         }
+
+        if(Gate::denies('owner', $reply->user->id))
+        {
+            abort(403, 'Not Authorized');
+        }
+
         return (bool) $reply->delete();
     }
 }
